stable state

+++ src/main/java/com/nikiforov/bank/controllers/ClientController.java
@@ -3,8 +3,11 @@ package com.nikiforov.bank.controllers;
 import com.nikiforov.bank.dto.Client;
 import com.nikiforov.bank.model.ClientEntity;
 import com.nikiforov.bank.services.ClientService;
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,12 +18,16 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 
 @RestController
 @RequestMapping("/client")
 @Data
+@NoArgsConstructor
 public class ClientController {
+    @Value("${feature.toggle.delete.client:false}")
+    private boolean featureToggleDeleteClient;
 
     @Autowired
     private ClientService clientService;
@@ -46,6 +53,10 @@ public class ClientController {
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     public void deleteClient(@PathVariable Long id) {
-        clientService.deleteClient(id);
+
+        if (this.featureToggleDeleteClient)
+            clientService.deleteClient(id);
+        else
+            throw new ResponseStatusException(HttpStatus.METHOD_NOT_ALLOWED, "Операция отключена администратором");
     }
 }
